Move _get_proposal function from data pipeline to dataset
